refactor: 프로젝트 구조에 맞게 컴포넌트 파일 이동 #106
